requirements.txt
django>=5.2
djangorestframework>=3.16.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.0
google-generativeai>=0.8.0